third_party/blink/web_tests/fast/dom/shadow/elementfrompoint-expected.txt

 
PASS shadowRoot.elementFromPoint() threw exception TypeError: Failed to execute 'elementFromPoint' on 'ShadowRoot': 2 arguments required, but only 0 present..
PASS shadowRoot.elementFromPoint(0) threw exception TypeError: Failed to execute 'elementFromPoint' on 'ShadowRoot': 2 arguments required, but only 1 present..
PASS shadowRoot.elementFromPoint(x12, y12) is box12
PASS nestedShadowRoot.elementFromPoint(x22, y22) is box22
PASS shadowRoot.elementFromPoint(x22, y22) is nestedHost
PASS document.elementFromPoint(x22, y22) is host
PASS root3.elementFromPoint(centerX(blockHost), centerY(blockHost)) is blockHost
PASS document.elementFromPoint(centerX(blockHost), centerY(blockHost)) is blockHost
PASS root4.elementFromPoint(centerX(inlineBlockHost), centerY(inlineBlockHost)) is inlineBlockHost
PASS document.elementFromPoint(centerX(inlineBlockHost), centerY(inlineBlockHost)) is inlineBlockHost
PASS document.elementFromPoint(centerX(submit), centerY(submit)) is submit
PASS successfullyParsed is true

TEST COMPLETE
